fix: update status delivery service로 이동
