+++ order-delivery/src/main/java/com/team12/order_delivery/delivery/service/DeliveryService.java
@@ -7,7 +7,11 @@ import com.team12.order_delivery.delivery.domain.Delivery;
 import com.team12.order_delivery.delivery.dto.DeliveryReqDto;
 import com.team12.order_delivery.delivery.dto.DeliveryResDto;
 import com.team12.order_delivery.delivery.repository.DeliveryRespository;
+import com.team12.order_delivery.deliveryRoute.domain.DeliveryRoute;
+import com.team12.order_delivery.deliveryRoute.dto.RouteResDto;
+import com.team12.order_delivery.deliveryRoute.repository.DeliveryRouteRepository;
 import com.team12.order_delivery.deliveryRoute.service.DeliveryRouteService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
@@ -21,6 +25,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class DeliveryService {
     private final DeliveryRespository deliveryRespository;
+    private final DeliveryRouteRepository deliveryRouteRepository;
     private final DeliveryRouteService deliveryRouteService;
 
     @Transactional
@@ -50,9 +55,9 @@ public class DeliveryService {
 
     public DeliveryResDto getDelivery(String deliveryId) {
         try {
-            return new DeliveryResDto(deliveryRespository.findById(UUID.fromString(deliveryId)).orElseThrow(() -> new BusinessLogicException(ExceptionCode.INVALID_PARAMETER)));
+            return new DeliveryResDto(deliveryRespository.findById(UUID.fromString(deliveryId)).orElseThrow(() ->
+                    new BusinessLogicException(ExceptionCode.INVALID_PARAMETER)));
         } catch (Exception e) {
-            log.info(e.getMessage());
             throw new BusinessLogicException(ExceptionCode.INVALID_PARAMETER);
         }
     }
@@ -88,15 +93,51 @@ public class DeliveryService {
         }
     }
 
+    @Transactional
     public DeliveryResDto updateDeliveryStatus(String deliveryId, String deliveryStatus) {
         try {
-            Delivery delivery = deliveryRespository.findById(UUID.fromString(deliveryId)).orElseThrow(() -> new IllegalArgumentException("배송 정보가 없습니다."));
+            Delivery delivery = deliveryRespository.findById(UUID.fromString(deliveryId))
+                    .orElseThrow(() -> new EntityNotFoundException("배송 정보가 없습니다."));
             delivery.setDeliveryStatus(Delivery.DeliveryStatus.valueOf(deliveryStatus));
             deliveryRespository.save(delivery);
             return new DeliveryResDto(delivery);
+        } catch (IllegalArgumentException e) {
+            log.error("Invalid status or UUID format: ", e);
+            throw new BusinessLogicException(ExceptionCode.INVALID_PARAMETER);
+        } catch (EntityNotFoundException e) {
+            log.error("Delivery not found: ", e);
+            throw new BusinessLogicException(ExceptionCode.ENTITY_NOT_FOUND);
+        } catch (Exception e) {
+            log.error("Unexpected error in updateDeliveryStatus: ", e);
+            return null;
+        }
+    }
+
+    @Transactional
+    public RouteResDto updateDeliveryRouteStatus(String deliveryRouteId, String deliveryRouteStatus) {
+        try {
+            DeliveryRoute deliveryRoute = deliveryRouteRepository.findById(UUID.fromString(deliveryRouteId))
+                    .orElseThrow(() -> new EntityNotFoundException("DeliveryRoute not found with id: " + deliveryRouteId));
+
+            DeliveryRoute.RouteStatus newStatus = DeliveryRoute.RouteStatus.valueOf(deliveryRouteStatus);
+
+            if(deliveryRoute.getSequence() == 1 && newStatus == DeliveryRoute.RouteStatus.DELIVERING) {
+                updateDeliveryStatus(String.valueOf(deliveryRoute.getDeliveryId()), "DELIVERING");
+            } else if (isLastRoute(deliveryRoute) && newStatus == DeliveryRoute.RouteStatus.ARRIVED) {
+                updateDeliveryStatus(String.valueOf(deliveryRoute.getDeliveryId()), "DELIVERED");
+            }
+
+            deliveryRoute.setStatus(newStatus);
+            deliveryRouteRepository.save(deliveryRoute);
+            return new RouteResDto(deliveryRoute);
         } catch (Exception e) {
             throw new BusinessLogicException(ExceptionCode.INVALID_PARAMETER);
         }
+    }
+
+    private boolean isLastRoute(DeliveryRoute route) {
+        int totalRoutes = deliveryRouteRepository.countByDeliveryId(route.getDeliveryId());
+        return totalRoutes == route.getSequence();
     }
 
 }
